+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * IdCardUtil提供hide方法，用于将固定位置的数字隐藏为星号
 * 增加parseBean重载（针对Dict和Entity）（pr#63@Github）
 * StrUtil增加replaceChars用于替换某些字符（感谢@【北京】宁静）
+* DateUtil.parse增加日期支持的格式，DatePattern增加纯数字形式日期（感谢@【北京】宁静）
 
 ### Bug修复
 * 修复MapUtil.filter方法修改值无效问题